Add authentication functionality (#33)
* creates dummy file for auth check

* 🛂 adds auth functionality to export endpoint

* 🎨 send response as json

* 🎨 changes response code from 403 to 401

* 🛂 adds auth functionality to import endpoint

* ♻️ removes uncecessary empty line

* 🛂 adds auth functionality to add_bookmark endpoint

* 🛂 adds auth functionality to delete_bookmark endpoint

* 🛂 adds auth functionality to update_bookmark endpoint

+++ src/controllers/deleteBookmarkController.js
@@ -1,5 +1,6 @@
 const readFile = require('../utils/readFile');
 const writeFile = require('../utils/writeFile');
+const authCheck = require('../utils/authCheck');
 
 function deleteChildNode(bookmarks, id_to_be_deleted) {
     for (let i = 0; i < bookmarks.length; i++) {
@@ -18,13 +19,18 @@ function deleteChildNode(bookmarks, id_to_be_deleted) {
 
 
 const handleDeleteBookmark = (req, res) => {
-    bookmarks = readFile.readBookmarksFile()
-    bookmarks = JSON.parse(bookmarks)
-    //TODO Add error handling
-    //result will be true or false if bookmark deleted or not
-    result = deleteChildNode(bookmarks, req.body.id)
-    writeFile.createBookmarksFile(JSON.stringify(result))
-    res.json({ message: 'Bookmark Deleted' });
+    if (authCheck.isAuthorized(req.headers.authorization)){
+        bookmarks = readFile.readBookmarksFile()
+        bookmarks = JSON.parse(bookmarks)
+        //TODO Add error handling
+        //result will be true or false if bookmark deleted or not
+        result = deleteChildNode(bookmarks, req.body.id)
+        writeFile.createBookmarksFile(JSON.stringify(result))
+        res.status(200).json({ message: 'Bookmark Deleted' });      
+    }
+    else{
+        res.status(401).json('Unauthorized');
+    }
 };
 
 module.exports = {
